Download a file Rest endpoint

+++ src/main/java/com/candidate/model/FileUploadRepository.java
@@ -8,5 +8,5 @@ import com.candidate.model.FileUploadMetaData;
  * A JPA repository used to perform crud operations on file meta data records in
  * database
  */
-public interface FileUploadRepository extends JpaRepository<FileUploadMetaData, Long> {
+public interface FileUploadRepository extends JpaRepository<FileUploadMetaData, Integer> {
 }
